move all the bbox functions for swap to bbox.h & bbox.c
