+++ src/file_storage/file_hdf5.c
@@ -6,6 +6,7 @@
 
 #include "dspaces-common.h"
 #include "util.h"
+#include "bbox.h"
 #include "file_storage/file_hdf5.h"
 
 #include <hdf5.h>
@@ -99,150 +100,6 @@ static void hdf5_dataset_name_parser(char* dname, struct bbox* bbox)
         }
     }
     
-}
-
-/*
-  Test if bounding boxes b0 and b1 intersect along dimension dim.
-*/
-static bool bbox_intersect_ondim(const struct bbox *b0, const struct bbox *b1,
-                                int dim)
-{
-    if((b0->lb.c[dim] <= b1->lb.c[dim] && b1->lb.c[dim] <= b0->ub.c[dim]) ||
-       (b1->lb.c[dim] <= b0->lb.c[dim] && b0->lb.c[dim] <= b1->ub.c[dim]))
-        return true;
-    else
-        return false;
-}
-
-/* Test if bounding box bbox0 includes bbox1 or bbox1 includes bbox0
- * along all dimensions. Return value could be:
- * 0 - no include relationship
- * 1 - bbox0 includes bbox1
- * 2 - bbox1 includes bbox0
- * */
-static int bbox_can_include(struct bbox* bbox0, struct bbox* bbox1)
-{
-    int ret;
-    struct bbox* big, *small;
-
-    // Find the bigger bbox first
-    if(bbox_dist(bbox0, 0)<=bbox_dist(bbox1, 0)) {
-        big = bbox1;
-        small = bbox0;
-        ret = 2;
-    } else {
-        big = bbox0;
-        small = bbox1;
-        ret = 1;
-    }
-
-    // If we cannot find a bbox whose size on every dimension is
-    // bigger than the other one, then there is no chance for either
-    // 1 bbox to include the other 
-    for(int i=1; i<bbox0->num_dims; i++) {
-        if(bbox_dist(small, i)>bbox_dist(big, i)) {
-            return 0;
-        }
-    }
-
-    for(int i=0; i<bbox0->num_dims; i++) {
-        if(!bbox_include_ondim(big, small, i)) {
-            return 0;
-        }
-    }
-
-    return ret;
-}
-
-/* Test if bounding box bbox0 and bbox1 can union.
- * Return value is the dimension where they can union along,
- * -1 when they cannot union.
- * */
-int bbox_can_union(const struct bbox* bbox0, const struct bbox* bbox1)
-{
-    bool share;
-    /* Two bbox can union means that they share a common
-     * (n-1) dim coordinates, and has 1 dimension extended. */
-    for(int i=0; i<bbox0->num_dims; i++) { // Choose 1 dimension
-        // check the coordinates of the other (n-1) dimension
-        share = true;
-        for(int j=0; j<bbox0->num_dims; j++) {
-            if(j==i) continue;
-            if(bbox0->lb.c[j] != bbox1->lb.c[j] || bbox0->ub.c[j] != bbox1->ub.c[j]) {
-                share = false;
-                break;
-            }
-        }
-
-        if(!share) continue; // choose the next dimension
-
-        if(share && bbox_intersect_ondim(bbox0, bbox1, i)) return i;
-    }
-    return -1;
-}
-
-static void bbox_union_ondim(const struct bbox* bbox0, const struct bbox* bbox1,
-                                int dim, struct bbox* bbox2)
-{
-    bbox2->num_dims = bbox0->num_dims;
-    for(int i=0; i<bbox0->num_dims; i++) {
-        if(i==dim) {
-            bbox2->lb.c[i] = min(bbox0->lb.c[i], bbox1->lb.c[i]);
-            bbox2->ub.c[i] = min(bbox0->ub.c[i], bbox1->ub.c[i]);
-        } else {
-            bbox2->lb.c[i] = bbox0->lb.c[i];
-            bbox2->ub.c[i] = bbox0->ub.c[i];
-        }
-    }
-}                                                
-
-/* Test if bounding box bbox0 and bbox1 can union and
- * compute their union bbox. The result will be stored in bbox2.
- * Return value is ture when they can union, otherwise false.
- * */
-static bool bbox_union(const struct bbox* bbox0, const struct bbox* bbox1, struct bbox* bbox2)
-{
-    bool found = false, share;
-    int dim;
-    /* Two bbox can union means that they share a common
-        (n-1) dim coordinates, and has 1 dimension extended. */
-    for(int i=0; i<bbox0->num_dims; i++) { // Choose 1 dimension
-        // check the coordinates of the other (n-1) dimension
-        share = true;
-        for(int j=0; j<bbox0->num_dims; j++) {
-            if(j==i) continue;
-            if(bbox0->lb.c[j] != bbox1->lb.c[j] || bbox0->ub.c[j] != bbox1->ub.c[j]) {
-                share = false;
-                break;
-            }
-        }
-
-        if(!share) continue; // choose the next dimension
-
-        found = share && bbox_intersect_ondim(bbox0, bbox1, i);
-
-        if(found) {
-            dim = i;
-            break;
-        }
-    }
-
-    if(!found) {
-        return false;
-    } else { // fill bbox2
-        bbox2->num_dims = bbox0->num_dims;
-        for(int i=0; i<bbox0->num_dims; i++) {
-            if(i==dim) {
-                bbox2->lb.c[i] = min(bbox0->lb.c[i], bbox1->lb.c[i]);
-                bbox2->ub.c[i] = min(bbox0->ub.c[i], bbox1->ub.c[i]);
-            } else {
-                bbox2->lb.c[i] = bbox0->lb.c[i];
-                bbox2->ub.c[i] = bbox0->ub.c[i];
-            }
-        }
-        return true;
-    }
-
 }
 
 /* Search the first dataset that can include the queried bbox from a HDF5 file.
